refactor: extract search paste dialog
